backend/requirements.txt
# ================================
# DEPENDENCIAS PRINCIPALES
# ================================

# Framework web
fastapi[security]==0.104.1
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Base de datos
sqlalchemy==2.0.23
greenlet==3.0.1

# Análisis de texto
textblob==0.17.1
vaderSentiment==3.3.2
nltk==3.8.1
deep-translator==1.11.4

# Utilidades
numpy==1.24.3
aiohttp==3.9.1
requests==2.31.0

# ================================
# MONITOREO Y LOGGING
# ================================
prometheus-client==0.19.0
psutil==5.9.6

# ================================
# SEGURIDAD Y AUTENTICACIÓN
# ================================

# JWT y tokens
python-jose[cryptography]==3.3.0

# Hashing de contraseñas
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Encriptación de datos
cryptography==41.0.7

# Validación de emails
email-validator==2.1.0

# ================================
# ANÁLISIS DE SEGURIDAD
# ================================
safety==2.3.5
bandit==1.7.5

# ================================
# GENERACIÓN DE REPORTES
# ================================
python-docx==1.1.0
lxml==5.1.0  # Dependencia de python-docx

# ================================
# TESTING Y QA
# ================================
pytest==7.4.3
pytest-html==4.1.1
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx==0.25.2  # Cliente HTTP asíncrono para testing